Updates in response to comments from referee #2.
The manuscript has been updated in accordance with the recommended
changes except for the following items:

1. Page 6: It might be sufficient to say > rather than Terminal> here
   (since Terminal is already mentioned on the top)

This is part of the template/doconce so we don't want to change it.

2. Page 7: Maybe move this URL to bitbuchet.com/fenicsproject?

We want to keep the repo url for historical and sentimental reasons.

3. Page 17: why not use xdmf? I thought pvd is outdated?

We should perhaps update in a future version when it is as stable
as pvd.

3. Page 21: ... belong to the boundary (which will be discussed later).

We don't understand this comment so we have skipped it for now.

4. Page 25: My personal opinion is that reusing the variable here adds
   to more confusion that readability.

Perhaps, but it would be too much work to change this consistently
throughout the book, maybe in the future.

5. Page 92: maybe show plot

We are already short on space for the 150 page book limit.

6. Page 92: this codeline is inconsistent with the text in the top of
   this page

We don't understand this comment so we have skipped it for now.

7. Page 92: no-op

We don't understand this comment so we have skipped it for now.

8. Page 92: maybe move the plot statement into the main function?

We want to keep it this way.

9. Page 116: It would be good to show a concrete example how to
   instiantate a mesh like this and plot it also.

Skipping this due to lack of space.

10. Page 118: mut he log level be set accordingly to see the output?

No.

11. Page 118: Why cant we just print(parameters)

We do have a __str__ method but it prints the non-verbose version of the
output so it will not be very informative.

12. Page 127: This global-local mix is a bit confusing

Yes, so we think it is a good idea to explain how it works... :-)

13. Page 134: why not use the (previously) mentioend error norm here?

We do use it, on the following page and explain its advantage.

14. Page 138: I dont see how this is optimistic?

We have updated the tables and the comments with correct numbers.

15. Page 142: does it have python3 support?

This is unclear, perhaps it will get it depending on who continues
to develop scitools.

16. Page 143: Parallelism is obviously an importaint missing part -
    but I guess you will be coverign this in the second part.

Yes!

17. Page 143: Overall a fantastic introduction to FEniCS - I will
    certaintly be making use of it in the future.

Thanks! :-)